Improved backend apis

+++ backend/controllers/portfolios.js
@@ -5,11 +5,40 @@ const fetch = require('node-fetch')
 const {StatusCodes} = require('http-status-codes')
 const {BadRequestError, UnauthenticatedError} = require('../errors');
 
+// Upon calling this API, a new entry will be inserted in portfolios
+// for a special stock called "dollars" along with the default
+// startingMoney amount.
 const createPortfolio = async (req, res) => {
-    req.body.userId = req.user.userId;
-    console.log(req.body)
-    const portfolio  = await Portfolio.create(req.body);
-    res.status(StatusCodes.CREATED).json({portfolio});
+    const startingMoney = 10000
+    const userID = req.user.userId;
+    // Default for now
+    const portID = 1
+
+    // First verify that the user has not already created max portfolios
+    const portfolioInfo = await Portfolio.find()
+    const portInfo = retrievePortInfoKernel(userID, portID, portfolioInfo)
+    if (portInfo.length > 0) {
+        // TEST CODE - Check for the presence of a "dollars" portfolio.
+        // This is being used only to correct past MongoDB entries
+        // Note that this only checks for one portfolio. It needs to check for 4.
+        const dollarsItem = portInfo.filter((item) => item.stockName === "dollars")
+        if (dollarsItem.length > 0){
+            res.status(400).json({success:false, msg:'You are already at max number of portfolios.'})
+        }
+        else{
+            // Make a new portfolio entry for "dollars".
+            portItem = ({userId:userID, portId:portID, stockName:"dollars", numOfUnits:startingMoney, initCost:1})
+            const portfolio = await Portfolio.create(portItem);
+            res.status(StatusCodes.CREATED).json({portfolio});
+        }
+    }
+    else
+    {
+        // Make a new portfolio entry for "dollars".
+        portItem = ({userId:userID, portId:portID, stockName:"dollars", numOfUnits:startingMoney, initCost:1})
+        const portfolio = await Portfolio.create(portItem);
+        res.status(StatusCodes.CREATED).json({portfolio});
+    }
 }
 
 // This is the API for the leaderboard and should be called whenever the
@@ -140,6 +169,8 @@ const getPortfolio = async (req, res) => {
 // We can adapt this later to accept whole dollar values and calculate
 // numOfUnits from that.
 const updatePortfolio = async (req, res) => {
+    let stockDoesntExist = false
+    let alreadyPurchased = false
     // const userID = req.params.id
     // const {portID, stockName, numOfUnits, initCost} = req.body
     let portfolioInfo = await Portfolio.find()
@@ -178,16 +209,35 @@ const updatePortfolio = async (req, res) => {
             if (stockName){
                 if (numOfUnits){
                     if (initCost){
-                        portItem = ({userId:userID, portId:portID, stockName:stockName, numOfUnits:numOfUnits, initCost:initCost})
-                        // Note that this method is pushing to the local instance of
-                        // the collection of portfolioData. It needs to be adapted to
-                        // push to MongoDB instead.
-                        await Portfolio.create(portItem);
+                        // First see if this stock exists. Note this will have to be checked
+                        // against the TwelveData API or cache later.
+                        currItem = currentData.filter((infoItem) => (infoItem.stockName===stockName))
+                        if (currItem.length == 0){
+                            stockDoesntExist = true
+                        }
+                        else {
+                            // Next check if the user already owns this stock. Implicit in this
+                            // is an "all or nothing" approach, in which a user may only sell or
+                            // purchase batches of one stock at at time.
+                            alreadyPurchased = await Portfolio.exists({userId:userID, portId:portID, stockName:stockName})
+                            if (alreadyPurchased){
+                                // Do nothing...exit the if statement
+                            }
+                            else{
+                                // Finally, before allowing the purchase to go through, verify
+                                // the user has the funds for it...
+                                portItem = ({userId:userID, portId:portID, stockName:stockName, numOfUnits:numOfUnits, initCost:initCost})
+                                // Note that this method is pushing to the local instance of
+                                // the collection of portfolioData. It needs to be adapted to
+                                // push to MongoDB instead.
+                                await Portfolio.create(portItem);
                         
-                        // portfolioInfo and currInfo is now stale...reload here.
-                        portfolioInfo = await Portfolio.find()
-                        currInfo = retrieveCurrInfoKernel(userID, portID, portfolioInfo)
-                        req.totalValue = totalValueKernel(userID,portID,portfolioInfo,currInfo)
+                                // portfolioInfo and currInfo is now stale...reload here.
+                                portfolioInfo = await Portfolio.find()
+                                currInfo = retrieveCurrInfoKernel(userID, portID, portfolioInfo)
+                                req.totalValue = totalValueKernel(userID,portID,portfolioInfo,currInfo)
+                            }
+                        }
                     }
                 }
             }
@@ -196,6 +246,12 @@ const updatePortfolio = async (req, res) => {
 
     if (req.totalValue){
         res.status(200).json({success:true, data:req.totalValue})
+    }
+    else if (stockDoesntExist){
+        res.status(400).json({success:false, msg:'The requested stock does not exist.'})
+    }
+    else if (alreadyPurchased){
+        res.status(400).json({success:false, msg:'You already own this stock. You must sell first.'})
     }
     else
     {
@@ -225,6 +281,7 @@ const sellPortfolioItem = async (req, res) => {
     const portfolioInfo = await Portfolio.find()
 
     let currInfo = []
+    let saleValue = 0
 
     const portInfo = retrievePortInfoKernel(userID, portID, portfolioInfo)
     currInfo = retrieveCurrInfoKernel(userID, portID, portfolioInfo)
@@ -255,24 +312,27 @@ const sellPortfolioItem = async (req, res) => {
                 // adapted to search MongoDB instead. This should only ever return
                 // one item.
                 stockItem = portfolioInfo.filter((portItem) => (String(portItem.userId)===String(userID) && Number(portItem.portId)==Number(portID) && String(portItem.stockName)===String(stockName)))
-                // Get the current price. This will need to come from the hosted
-                // stock API. Again, this should only ever return one item.
-                //const twelveDataRes = await fetch(`https://api.twelvedata.com/price?symbol=${stockItem[0].stockName}&apikey=<YOURAPIKEY>`)
-                //const priceItem = await twelveDataRes.json()
-                currItem = currInfo.filter((infoItem) => (infoItem.stockName===stockName))
-                // Compute the sale value.
-                //req.saleValue = (stockItem[0].numOfUnits * Number(priceItem.price))
-                req.saleValue = (stockItem[0].numOfUnits * Number(currItem[0].currentCost))
-                // Finally, remove the sold item. This is temporary here, but will
-                // be modifying MongoDB in the actual version.
-                //portfolioData = portfolioData.filter((portItem) => (portItem.userID!=userID || portItem.portID!=portID || portItem.stockName!==stockName))
-                await Portfolio.deleteOne({userId: stockItem[0].userId, portId: stockItem[0].portId, stockName: stockItem[0].stockName})
+                // If we return nothing, the item doesn't exist so don't try and sell.
+                if (stockItem.length > 0){
+                    // Get the current price. This will need to come from the hosted
+                    // stock API. Again, this should only ever return one item.
+                    //const twelveDataRes = await fetch(`https://api.twelvedata.com/price?symbol=${stockItem[0].stockName}&apikey=<YOURAPIKEY>`)
+                    //const priceItem = await twelveDataRes.json()
+                    currItem = currInfo.filter((infoItem) => (infoItem.stockName===stockName))
+                    // Compute the sale value.
+                    //req.saleValue = (stockItem[0].numOfUnits * Number(priceItem.price))
+                    saleValue = (stockItem[0].numOfUnits * Number(currItem[0].currentCost))
+                    // Finally, remove the sold item. This is temporary here, but will
+                    // be modifying MongoDB in the actual version.
+                    //portfolioData = portfolioData.filter((portItem) => (portItem.userID!=userID || portItem.portID!=portID || portItem.stockName!==stockName))
+                    await Portfolio.deleteOne({userId: stockItem[0].userId, portId: stockItem[0].portId, stockName: stockItem[0].stockName})
+                }
             }
         }
     }
     
-    if (req.saleValue){
-        res.status(200).json({success:true, data:req.saleValue})
+    if (saleValue){
+        res.status(200).json({success:true, data:saleValue})
     }
     else
     {
@@ -395,7 +455,7 @@ function stockReturnKernel(userID, portID, portfolioInfo, currInfo){
     let returnCollection = []
     for (pI of portInfo){
         // Should only ever return one item
-        currentItem = currInfo.filter((infoItem) => {
+        currentItem = currentData.filter((infoItem) => {
             if (pI.stockName === infoItem.stockName){
                 return infoItem
             }
